+++ modules/channel-vonage/src/renderers/index.ts
@@ -3,5 +3,6 @@ export * from './card'
 export * from './carousel'
 export * from './choices'
 export * from './image'
+export * from './template'
 export * from './text'
 export * from './video'
